task_management_api/requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==5.1.4
django-environ==0.11.2
django-filter==24.3
django-heroku==0.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
environ==1.0
gunicorn==20.1.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
PyJWT==2.10.1
setuptools==75.7.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
whitenoise==6.8.2
